some validations

+++ Avalon.Client/SearchEditBeerWin.xaml.cs
@@ -1,18 +1,6 @@
 ﻿using Avalon.Models;
 using Avalon.Service;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Shapes;
 
 namespace Avalon.Client
 {
@@ -28,30 +16,30 @@ namespace Avalon.Client
 
         }
 
-
         private void CancelButton_Click(object sender, RoutedEventArgs e)
         {
             this.Close();
-
         }
 
         private void EditButton_Click(object sender, RoutedEventArgs e)
         {
-            EditBeerWin editBeerWin = new EditBeerWin(this.beersDatagrid.SelectedItem as Beer);
-            this.Close();
-            editBeerWin.Show();
-            
-
+            if(this.beersDatagrid.SelectedItem != null)
+            {
+                EditBeerWin editBeerWin = new EditBeerWin(this.beersDatagrid.SelectedItem as Beer);
+                this.Close();
+                editBeerWin.Show();
+            }
         }
 
         private void DeleteButton_Click(object sender, RoutedEventArgs e)
         {
-            string beerName = (this.beersDatagrid.SelectedItem as Beer).Name;
+            if(this.beersDatagrid.SelectedItem != null)
+            {
+                string beerName = (this.beersDatagrid.SelectedItem as Beer).Name;
 
-            BeerService.DeleteBeer(beerName);
-            beersDatagrid.DataContext = BeerService.GetAllBeers();
-
-
+                BeerService.DeleteBeer(beerName);
+                beersDatagrid.DataContext = BeerService.GetAllBeers();
+            }
         }
 
 
@@ -67,7 +55,6 @@ namespace Avalon.Client
             }
 
             beersDatagrid.DataContext = BeerService.GetBeersByName(beerName);
-
         }
     }
 }
